- now compiles with -ansi -pedantic
git-svn-id: https://trac.beirdo.ca:4443/svn/havokmud/branches/rev_1_11@853 69fa473d-82f3-0310-a870-f1ebe01cfa2b

+++ havokmud/mind_use1.c
@@ -10,7 +10,7 @@
 
 extern char    *dirs[];
 
-void mind_use_burn(byte level, struct char_data *ch, char *arg, int type,
+void mind_use_burn(int level, struct char_data *ch, char *arg, int type,
                    struct char_data *victim, struct obj_data *tar_obj)
 {
     switch (type) {
@@ -26,7 +26,7 @@ void mind_use_burn(byte level, struct char_data *ch, char *arg, int type,
     }
 }
 
-void mind_use_teleport(byte level, struct char_data *ch, char *arg,
+void mind_use_teleport(int level, struct char_data *ch, char *arg,
                        int type, struct char_data *victim,
                        struct obj_data *tar_obj)
 {
@@ -43,7 +43,7 @@ void mind_use_teleport(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_probability_travel(byte level, struct char_data *ch,
+void mind_use_probability_travel(int level, struct char_data *ch,
                                  char *arg, int type,
                                  struct char_data *victim,
                                  struct obj_data *tar_obj)
@@ -61,7 +61,7 @@ void mind_use_probability_travel(byte level, struct char_data *ch,
     }
 }
 
-void mind_use_danger_sense(byte level, struct char_data *ch, char *arg,
+void mind_use_danger_sense(int level, struct char_data *ch, char *arg,
                            int type, struct char_data *victim,
                            struct obj_data *tar_obj)
 {
@@ -78,7 +78,7 @@ void mind_use_danger_sense(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_clairvoyance(byte level, struct char_data *ch, char *arg,
+void mind_use_clairvoyance(int level, struct char_data *ch, char *arg,
                            int type, struct char_data *victim,
                            struct obj_data *tar_obj)
 {
@@ -95,7 +95,7 @@ void mind_use_clairvoyance(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_disintegrate(byte level, struct char_data *ch, char *arg,
+void mind_use_disintegrate(int level, struct char_data *ch, char *arg,
                            int type, struct char_data *victim,
                            struct obj_data *tar_obj)
 {
@@ -112,7 +112,7 @@ void mind_use_disintegrate(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_telekinesis(byte level, struct char_data *ch, char *arg,
+void mind_use_telekinesis(int level, struct char_data *ch, char *arg,
                           int type, struct char_data *victim,
                           struct obj_data *tar_obj)
 {
@@ -160,7 +160,7 @@ void mind_use_telekinesis(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_levitation(byte level, struct char_data *ch, char *arg,
+void mind_use_levitation(int level, struct char_data *ch, char *arg,
                          int type, struct char_data *victim,
                          struct obj_data *tar_obj)
 {
@@ -177,7 +177,7 @@ void mind_use_levitation(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_cell_adjustment(byte level, struct char_data *ch, char *arg,
+void mind_use_cell_adjustment(int level, struct char_data *ch, char *arg,
                               int type, struct char_data *victim,
                               struct obj_data *tar_obj)
 {
@@ -194,7 +194,7 @@ void mind_use_cell_adjustment(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_chameleon(byte level, struct char_data *ch, char *arg,
+void mind_use_chameleon(int level, struct char_data *ch, char *arg,
                         int type, struct char_data *victim,
                         struct obj_data *tar_obj)
 {
@@ -211,7 +211,7 @@ void mind_use_chameleon(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_psi_strength(byte level, struct char_data *ch, char *arg,
+void mind_use_psi_strength(int level, struct char_data *ch, char *arg,
                            int type, struct char_data *victim,
                            struct obj_data *tar_obj)
 {
@@ -228,7 +228,7 @@ void mind_use_psi_strength(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_mind_over_body(byte level, struct char_data *ch, char *arg,
+void mind_use_mind_over_body(int level, struct char_data *ch, char *arg,
                              int type, struct char_data *victim,
                              struct obj_data *tar_obj)
 {
@@ -245,7 +245,7 @@ void mind_use_mind_over_body(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_domination(byte level, struct char_data *ch, char *arg,
+void mind_use_domination(int level, struct char_data *ch, char *arg,
                          int type, struct char_data *victim,
                          struct obj_data *tar_obj)
 {
@@ -262,7 +262,7 @@ void mind_use_domination(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_mind_wipe(byte level, struct char_data *ch, char *arg,
+void mind_use_mind_wipe(int level, struct char_data *ch, char *arg,
                         int type, struct char_data *victim,
                         struct obj_data *tar_obj)
 {
@@ -279,7 +279,7 @@ void mind_use_mind_wipe(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_psychic_crush(byte level, struct char_data *ch, char *arg,
+void mind_use_psychic_crush(int level, struct char_data *ch, char *arg,
                             int type, struct char_data *victim,
                             struct obj_data *tar_obj)
 {
@@ -296,7 +296,7 @@ void mind_use_psychic_crush(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_tower_iron_will(byte level, struct char_data *ch, char *arg,
+void mind_use_tower_iron_will(int level, struct char_data *ch, char *arg,
                               int type, struct char_data *victim,
                               struct obj_data *tar_obj)
 {
@@ -313,7 +313,7 @@ void mind_use_tower_iron_will(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_mindblank(byte level, struct char_data *ch, char *arg,
+void mind_use_mindblank(int level, struct char_data *ch, char *arg,
                         int type, struct char_data *victim,
                         struct obj_data *tar_obj)
 {
@@ -330,7 +330,7 @@ void mind_use_mindblank(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_psychic_impersonation(byte level, struct char_data *ch,
+void mind_use_psychic_impersonation(int level, struct char_data *ch,
                                     char *arg, int type,
                                     struct char_data *victim,
                                     struct obj_data *tar_obj)
@@ -348,7 +348,7 @@ void mind_use_psychic_impersonation(byte level, struct char_data *ch,
     }
 }
 
-void mind_use_ultra_blast(byte level, struct char_data *ch, char *arg,
+void mind_use_ultra_blast(int level, struct char_data *ch, char *arg,
                           int type, struct char_data *victim,
                           struct obj_data *tar_obj)
 {
@@ -365,7 +365,7 @@ void mind_use_ultra_blast(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_intensify(byte level, struct char_data *ch, char *arg,
+void mind_use_intensify(int level, struct char_data *ch, char *arg,
                         int type, struct char_data *victim,
                         struct obj_data *tar_obj)
 {
@@ -382,7 +382,7 @@ void mind_use_intensify(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_mind_tap(byte level, struct char_data *ch, char *arg,
+void mind_use_mind_tap(int level, struct char_data *ch, char *arg,
                        int type, struct char_data *victim,
                        struct obj_data *tar_obj)
 {
@@ -399,7 +399,7 @@ void mind_use_mind_tap(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_kinolock(byte level, struct char_data *ch, char *arg,
+void mind_use_kinolock(int level, struct char_data *ch, char *arg,
                        int type, struct char_data *tar_ch,
                        struct obj_data *tar_obj)
 {
@@ -416,7 +416,7 @@ void mind_use_kinolock(byte level, struct char_data *ch, char *arg,
     }
 }
 
-void mind_use_sense_object(byte level, struct char_data *ch, char *arg,
+void mind_use_sense_object(int level, struct char_data *ch, char *arg,
                            int si, int type, struct char_data *tar_ch,
                            struct obj_data *tar_obj)
 {
