+++ havokmud/spells.h
@@ -597,10 +597,10 @@
 #define TAR_GROUP        (1<<14)        /* spells that target groupees */
 
 struct spell_info_type {
-    void            (*spell_pointer)(byte level, struct char_data * ch, 
-		                     char *arg, int type, 
-				     struct char_data * tar_ch,
-				     struct obj_data * tar_obj);
+    void            (*spell_pointer)(int level, struct char_data * ch, 
+                                     char *arg, int type, 
+                                     struct char_data * tar_ch,
+                                     struct obj_data * tar_obj);
     byte            minimum_position;   /* Position for caster */
     ubyte           min_usesmana;       /* Amount of mana used by a spell */
     byte            beats;      /* Heartbeats until ready for next */
